model/requirements.txt
fastapi
uvicorn
rembg
pillow
python-multipart
pyjwt
onnxruntime